Replace apm with cat command
"apm -l" is a utility part of the OpenBSD base system and is available on linux but requires kernel changes and something like acpi is more common. Using this cat command however, works on most linux system without installing any dependencies although it is calling for "BAT0" which is the most common battery name but this may cause errors - especially with computers that don't have batteries.